New App Wizard Introduced
This Commit introduces the new application wizard to register an owner.

To do so, it creates the new wizard component, an unathenticated route to check if an owner account exists in the system to determine when to load it.

Some changes were also made to the BackendReducer and Authentication redux to accomodate the loading of this information and methods to regisater the first owner.

I also reworked the app.js component so that it will be much easier in the future to manage security.
