+++ Tombolo/client-reactjs/src/components/common/Constants.js
@@ -54,6 +54,7 @@ export const Constants = {
   PROPAGATIONS_CHANGES_ERROR: 'PROPAGATIONS_CHANGES_ERROR',
 
   SET_BACKEND_STATUS: 'SET_BACKEND_STATUS',
+  SET_USER_STATUS: 'SET_USER_STATUS',
 
   PROPAGATIONS_CURRENT_INITIATE: 'PROPAGATIONS_CURRENT_INITIATE',
   PROPAGATIONS_CURRENT_SUCCESS: 'PROPAGATIONS_CURRENT_SUCCESS',
