+++ Tombolo/client-reactjs/src/redux/reducers/BackendReducer.js
@@ -4,6 +4,8 @@ import { Constants } from '../../components/common/Constants';
 const initialState = {
   isConnected: false,
   statusRetrieved: false,
+  ownerExists: false,
+  ownerRetrieved: false,
 };
 
 // Backend reducer
@@ -14,6 +16,12 @@ export function backendReducer(state = initialState, action) {
         ...state,
         isConnected: action.payload,
         statusRetrieved: true,
+      };
+    case Constants.SET_USER_STATUS:
+      return {
+        ...state,
+        ownerExists: action.payload,
+        ownerRetrieved: true,
       };
     default:
       return state;
