--- Tombolo/server/routes/status/read.js
@@ -1,9 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-//route just to check if backend is running
-router.get("/", (req, res) => {
-  res.send("Tombolo's Backend is running succesfully");
-});
-
-module.exports = router;
